Refactored the whole codebase, both back-end and front-end and added controller for all api endpoints.

+++ backend/OnlineLibrary/src/main/java/com/arka99/OnlineLibrary/dto/requests/CheckoutBookRequest.java
@@ -4,6 +4,6 @@ import jakarta.validation.constraints.NotNull;
 
 public record CheckoutBookRequest(
     @NotNull
-    Integer bookId
+    Long bookId
 ) {
 }
